move common auth classes to be reused for AVS
