+++ common/src/main/java/module-info.java
@@ -1,9 +1,9 @@
 module org.example.age.common {
+    exports org.example.age.common.auth;
     exports org.example.age.common.account;
     exports org.example.age.common.html;
     exports org.example.age.common.server;
     exports org.example.age.common.site.api;
-    exports org.example.age.common.site.auth;
     exports org.example.age.common.site.config;
     exports org.example.age.common.site.verification;
