+++ common/src/main/java/org/example/age/common/auth/UserAgentAuthMatchDataExtractor.java
@@ -1,4 +1,4 @@
-package org.example.age.common.site.auth;
+package org.example.age.common.auth;
 
 import io.undertow.server.HttpServerExchange;
 import io.undertow.util.Headers;
