+++ common/src/main/java/org/example/age/common/auth/UserAgentAuthMatchDataExtractorModule.java
@@ -1,4 +1,4 @@
-package org.example.age.common.site.auth;
+package org.example.age.common.auth;
 
 import dagger.Binds;
 import dagger.Module;
