+++ common/src/main/java/org/example/age/common/site/api/SiteApiModule.java
@@ -5,8 +5,8 @@ import dagger.Module;
 import io.undertow.server.HttpHandler;
 import javax.inject.Named;
 import org.example.age.common.account.AccountIdExtractor;
+import org.example.age.common.auth.AuthMatchDataExtractor;
 import org.example.age.common.client.internal.RequestDispatcherModule;
-import org.example.age.common.site.auth.AuthMatchDataExtractor;
 import org.example.age.common.site.auth.internal.AuthManagerModule;
 import org.example.age.common.site.config.SiteConfig;
 import org.example.age.common.site.config.internal.SiteConfigurerModule;
