+++ common/src/main/java/org/example/age/common/site/auth/internal/AuthManagerImpl.java
@@ -5,8 +5,8 @@ import io.undertow.util.StatusCodes;
 import java.util.Optional;
 import javax.inject.Inject;
 import javax.inject.Singleton;
-import org.example.age.common.site.auth.AuthMatchData;
-import org.example.age.common.site.auth.AuthMatchDataExtractor;
+import org.example.age.common.auth.AuthMatchData;
+import org.example.age.common.auth.AuthMatchDataExtractor;
 import org.example.age.common.store.internal.PendingStore;
 import org.example.age.data.SecureId;
 import org.example.age.data.certificate.AgeCertificate;
