+++ common/src/main/java/org/example/age/common/site/auth/internal/AuthManagerModule.java
@@ -2,7 +2,7 @@ package org.example.age.common.site.auth.internal;
 
 import dagger.Binds;
 import dagger.Module;
-import org.example.age.common.site.auth.AuthMatchDataExtractor;
+import org.example.age.common.auth.AuthMatchDataExtractor;
 
 /**
  * Dagger module that publishes a binding for {@link AuthManager}.
