+++ common/src/test/java/org/example/age/common/auth/DisabledAuthMatchDataExtractorTest.java
@@ -1,4 +1,4 @@
-package org.example.age.common.site.auth;
+package org.example.age.common.auth;
 
 import static org.mockito.Mockito.mock;
