+++ common/src/test/java/org/example/age/common/site/api/testing/FakeAvsHandler.java
@@ -17,8 +17,8 @@ import javax.inject.Singleton;
 import okhttp3.HttpUrl;
 import okhttp3.Request;
 import okhttp3.RequestBody;
+import org.example.age.common.auth.AuthMatchDataExtractor;
 import org.example.age.common.client.internal.RequestDispatcher;
-import org.example.age.common.site.auth.AuthMatchDataExtractor;
 import org.example.age.data.SecureId;
 import org.example.age.data.VerifiedUser;
 import org.example.age.data.certificate.AgeCertificate;
