+++ common/src/test/java/org/example/age/common/site/auth/internal/AuthManagerTest.java
@@ -10,9 +10,9 @@ import io.undertow.util.StatusCodes;
 import java.time.Duration;
 import java.util.Map;
 import javax.inject.Singleton;
-import org.example.age.common.site.auth.AuthMatchData;
-import org.example.age.common.site.auth.AuthMatchDataExtractor;
-import org.example.age.common.site.auth.UserAgentAuthMatchDataExtractorModule;
+import org.example.age.common.auth.AuthMatchData;
+import org.example.age.common.auth.AuthMatchDataExtractor;
+import org.example.age.common.auth.UserAgentAuthMatchDataExtractorModule;
 import org.example.age.data.SecureId;
 import org.example.age.data.VerifiedUser;
 import org.example.age.data.certificate.AgeCertificate;
